+++ src/modules/mouse/Mouse.h
@@ -62,12 +62,12 @@ public:
 
 	virtual bool hasCursor() const = 0;
 
-	virtual int getX() const = 0;
-	virtual int getY() const = 0;
-	virtual void getPosition(int &x, int &y) const = 0;
-	virtual void setX(int x) = 0;
-	virtual void setY(int y) = 0;
-	virtual void setPosition(int x, int y) = 0;
+	virtual double getX() const = 0;
+	virtual double getY() const = 0;
+	virtual void getPosition(double &x, double &y) const = 0;
+	virtual void setX(double x) = 0;
+	virtual void setY(double y) = 0;
+	virtual void setPosition(double x, double y) = 0;
 	virtual void setVisible(bool visible) = 0;
 	virtual bool isDown(Button *buttonlist) const = 0;
 	virtual bool isVisible() const = 0;
